api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
supabase==2.7.4
python-dotenv==1.0.1
pydantic==2.8.2
anyio==3.7.1
httpx==0.27.0
google-generativeai>=0.8.0
